pyproject.toml
[tool.poetry]
name = "llm-libreco"
version = "0.1.0"
description = ""
authors = ["ingeniom22 <jamesmichael0444@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<3.12"
librecommender = "^1.5.1"
langchain = "^0.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
